+++ jreader/reader_array_test.go
@@ -11,14 +11,14 @@ func TestAddErrorStopsArrayParsing(t *testing.T) {
 	r := NewReader([]byte("[1,2]"))
 	arr := r.Array()
 	require.True(t, arr.Next())
-	require.Equal(t, 1, r.Int())
+	require.Equal(t, int64(1), r.Int64())
 
 	err := errors.New("sorry")
 	r.AddError(err)
 	require.Equal(t, err, r.Error())
 
 	require.False(t, arr.Next())
-	require.Equal(t, 0, r.Int())
+	require.Equal(t, int64(0), r.Int64())
 	require.Equal(t, err, r.Error())
 }
 
@@ -26,6 +26,6 @@ func TestSyntaxErrorStopsArrayParsing(t *testing.T) {
 	r := NewReader([]byte("[bad,1,2]"))
 	arr := r.Array()
 	require.False(t, arr.Next())
-	require.Equal(t, 0, r.Int())
+	require.Equal(t, int64(0), r.Int64())
 	require.Error(t, r.Error())
 }
